+++ data_linter/validators/__init__.py
@@ -1,3 +1,1 @@
 from .pandas_validator import PandasValidator  # noqa
-from .frictionless_validator import FrictionlessValidator  # noqa
-from .great_expectations_validator import GreatExpectationsValidator  # noqa